requirements.txt
altgraph==0.17
matplotlib==3.4.2
matplotlib-inline==0.1.2
nbconvert==6.0.7
nbformat==5.1.3
numpy==1.19.5
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
Pygments==2.9.0
scipy==1.6.3
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tk==0.1.0
Werkzeug==2.0.1
